+++ src/theme/components/header/header.ts
@@ -5,6 +5,7 @@ export const headerFocusColor = "#0F465C";
 export const headerLightBlue = "#78CCED";
 export const headerLightBlueIcon = "#1DA1D4";
 export const headerRed = "#ED1C24";
+export const headerRedDarkMode = "#DC6F73";
 export const headerRedDonate = "#E32B31";
 export const headerYellow = "#FEE34A";
 export const headerFocus = {
@@ -19,6 +20,9 @@ const Header = {
   parts: ["container", "horizontalRule", "logo"],
   baseStyle: {
     fontFamily: "'system-ui', 'Segoe UI', Tahoma, 'Helvetica', 'arial'",
+    _dark: {
+      backgroundColor: "dark.ui.bg.default",
+    },
     container: {
       marginX: { md: "20px", xl: "auto" },
       maxWidth: "1300px",
@@ -28,6 +32,9 @@ const Header = {
       bg: headerRed,
       marginTop: "0",
       marginBottom: "0",
+      _dark: {
+        backgroundColor: "dark.brand.primary",
+      },
     },
     logo: {
       padding: { base: "0 10px", md: "0" },