+++ src/theme/components/heading.ts
@@ -73,6 +73,9 @@ const Heading = {
       : isLowercase
       ? "lowercase"
       : null,
+    _dark: {
+      color: "dark.ui.typography.heading",
+    },
   }),
   // Available variants:
   // h1, h2, h3, h4, h5, h6,