+++ src/theme/components/helperErrorText.ts
@@ -7,7 +7,10 @@ const helperErrorText = {
     marginTop: "xxs",
     marginBottom: "0",
     fontSize: "helper.default",
-    color: isInvalid ? "ui.error.primary" : "ui.black",
+    color: isInvalid ? "ui.error.primary" : null,
+    _dark: {
+      color: isInvalid ? "dark.ui.error.primary" : null,
+    },
   }),
 };