+++ src/theme/components/skeletonLoader.ts
@@ -13,8 +13,11 @@ const element = {
 };
 const borderRules = {
   border: "1px solid",
-  borderColor: "ui.gray.light-cool",
+  borderColor: "ui.gray.x-light-cool",
   padding: "s",
+  _dark: {
+    borderColor: "ui.gray.xx-dark",
+  },
 };
 const imagePaddingBottomStyles = {
   landscape: "50%",
@@ -105,8 +108,11 @@ const fade = () =>
 const Skeleton = {
   baseStyle: {
     borderRadius: "2px",
-    background: "ui.gray.light-cool",
+    bg: "ui.gray.light-cool",
     animation: `${fade()} 1000ms cubic-bezier(0.25, -0.5, 1, 0) infinite`,
+    _dark: {
+      bg: "dark.ui.bg.hover",
+    },
   },
 };