+++ src/theme/components/statusBadge.ts
@@ -9,6 +9,12 @@ const StatusBadge = {
     px: "inset.narrow",
     textAlign: "center",
     width: "fit-content",
+    _dark: {
+      backgroundColor: "dark.ui.bg.active",
+      borderLeft: "4px solid",
+      borderColor: "ui.gray.medium",
+      color: "dark.ui.typography.heading",
+    },
   },
   variants: {
     low: {
@@ -16,9 +22,15 @@ const StatusBadge = {
     },
     medium: {
       bg: "ui.status.primary",
+      _dark: {
+        borderColor: "ui.status.primary",
+      },
     },
     high: {
       bg: "ui.status.secondary",
+      _dark: {
+        borderColor: "dark.ui.error.primary",
+      },
     },
   },
 };